Changed services to Transient Services.
Accoring to https://docs.microsoft.com/en-us/aspnet/core/fundamentals/dependency-injection?view=aspnetcore-3.0#service-lifetimes
Transient works best for "lightweight, stateless services", which these are.
And if official documentation says so, I am willing to trust it on this.

I originally set them to singletons as a pre-mature optimization since I figured that state-less services can be reused over and over so no need to re-instatiate them. But it seems, like regular singletons, that this is a bit of a anti-pattern.