Added empty alt attribute to images.
Don't really like to just say there is no alt-text.
But restructuring code to allow injecting the text is not on the schedule right now.
And the only text I can resonably inject there is the prize short description.
So any one visiting the site with a screen reader or other issues viewing the images can instead read the description.